Move harvest into its own module.
